+++ src/main/java/com/br/gcm/controller/EmpresaController.java
@@ -6,8 +6,10 @@ import com.br.gcm.dao.PaisDao;
 import com.br.gcm.dao.UfDao;
 import com.br.gcm.model.Empresa;
 import com.br.gcm.model.MensagemTransacao;
+import com.br.gcm.model.Usuario;
 import com.br.gcm.service.EmpresaService;
 import com.br.gcm.tag.Pagina;
+import com.br.gcm.util.Rotinas;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
@@ -40,47 +42,53 @@ public class EmpresaController {
     @Inject private PaisDao paisDao;
     @Inject private UfDao ufDao;
     @Inject private EmpresaService empresaService;
-
-    private String mensagem = "";
-    private int tipo = 9;
-
-    private  void limparmensagem(){
-        mensagem = "";
-        tipo = 9;
-    }
+    @Inject private Rotinas rotinas;
 
     @InitBinder
     public void initBinder(WebDataBinder binder) {
-        SimpleDateFormat dateFormat = new SimpleDateFormat("dd/MM/yyyy");
+        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
         dateFormat.setLenient(false);
         binder.registerCustomEditor(Date.class, new CustomDateEditor(dateFormat, false));
     }
 
     @RequestMapping(value = "/empresa_lista")
     public String lista(@PageableDefault(size = 10) Pageable pageable, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "1004");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+        Boolean novo     = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100401");
+        Boolean editar   = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100402");
+        Boolean deletar  = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100403");
+        Boolean detalhes = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100404");
 
-        MensagemTransacao mensagemTransacao = new MensagemTransacao();
-        mensagemTransacao.setTipo(tipo);
-        mensagemTransacao.setMensagem(mensagem);
-
-        model.addAttribute("mensagem", mensagemTransacao);
+        model.addAttribute("novo", novo);
+        model.addAttribute("editar", editar);
+        model.addAttribute("deletar", deletar);
+        model.addAttribute("detalhes", detalhes);
         model.addAttribute("empresa_lista", empresaDao.selectAll_paginado(pageable));
         model.addAttribute("pagina", new Pagina(pageable, empresaDao.count()));
 
-        limparmensagem();
         return "empresa_lista";
     }
 
-    //Deleta municipio
+    //Deleta
     @RequestMapping(value = "/empresa_deleta/{id_empresa}")
-    public String deletar(@PathVariable("id_empresa") Integer id_empresa) {
+    public String deletar(@PathVariable("id_empresa") Integer id_empresa, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100403");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         try{
             empresaService.delete(id_empresa);
-            tipo = 0;
-            mensagem = "Registro deletado com sucesso.";
         }catch(Exception e){
-            tipo = 1;
-            mensagem = e.getCause().toString();
+            model.addAttribute("mensagem", e.getCause().getMessage().toString());
+            return "mensagemerro";
         }
         return "redirect:/empresa_lista";
     }
@@ -88,6 +96,13 @@ public class EmpresaController {
     //Nova
     @RequestMapping(value = "/empresa_novo", method = RequestMethod.GET)
     public String novo(ModelMap model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100401");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         Empresa empresa = new Empresa();
         model.addAttribute("empresa", empresa);
         model.addAttribute("lista_pais", paisDao.Pais_lista());
@@ -96,25 +111,27 @@ public class EmpresaController {
 
     //Insert
     @RequestMapping(value = "/empresa_insert", method = RequestMethod.POST)
-    public ModelAndView insert(@ModelAttribute Empresa empresa, @PageableDefault(size = 10) Pageable pageable, BindingResult result) {
+    public String insert(@ModelAttribute Empresa empresa, Model model) {
         try{
             empresaService.insert(empresa);
-            tipo = 0;
-            mensagem = "Registro Inserido com sucesso.";
         }catch(Exception e){
-            tipo = 1;
-            mensagem = e.getCause().toString();
+            model.addAttribute("mensagem", e.getCause().getMessage().toString());
+            return "mensagemerro";
         }
-        ModelAndView mav = new ModelAndView();
-        mav.addObject("lista", empresaDao.selectAll_paginado(pageable));
-        mav.addObject("pagina", new Pagina(pageable, empresaDao.count()));
-        mav.setViewName("redirect:/empresa_lista");
-        return mav;
+
+        return "redirect:/empresa_lista";
     }
 
     //Editar
     @RequestMapping(value = "/empresa_editar/{id}", method = RequestMethod.GET)
     public String editar(@PathVariable("id") Integer id, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100402");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
+
         Empresa empresa = empresaDao.selectById(id);
         model.addAttribute("empresa", empresa);
         model.addAttribute("lista_pais", paisDao.Pais_lista());
@@ -123,25 +140,26 @@ public class EmpresaController {
 
     //Update
     @RequestMapping(value = "/empresa_update", method = RequestMethod.POST)
-    public ModelAndView update(@ModelAttribute Empresa empresa, @PageableDefault(size = 10) Pageable pageable, BindingResult result) {
+    public String update(@ModelAttribute Empresa empresa, Model model) {
         try{
             empresaService.update(empresa);
-            tipo = 0;
-            mensagem = "Registro Alterado com sucesso.";
         }catch(Exception e){
-            tipo = 1;
-            mensagem = e.getCause().toString();
+            model.addAttribute("mensagem", e.getCause().getMessage().toString());
+            return "mensagemerro";
         }
-        ModelAndView mav = new ModelAndView();
-        mav.addObject("lista", empresaDao.selectAll_paginado(pageable));
-        mav.addObject("pagina", new Pagina(pageable, empresaDao.count()));
-        mav.setViewName("redirect:/empresa_lista");
-        return mav;
+
+        return "redirect:/empresa_lista";
     }
 
     //Editar
     @RequestMapping(value = "/empresa_detalhes/{id}", method = RequestMethod.GET)
     public String detalhes(@PathVariable("id") Integer id, Model model) {
+        Usuario usuario = rotinas.usuarioLogado();
+        Boolean lista = rotinas.validaTransacaoUsuario(usuario.getId_usuario(), "100404");
+        if (lista != true) {
+            model.addAttribute("mensagem", "AVISO: Transação não permitida.");
+            return "mensagemerro";
+        }
         Empresa empresa = empresaDao.selectById(id);
         model.addAttribute("empresa", empresa);
         model.addAttribute("lista_pais", paisDao.Pais_lista());
